ai-service/requirements.txt
# Core FastAPI and server
fastapi==0.117.1
uvicorn==0.37.0
python-multipart==0.0.20
python-dotenv==1.1.1
httpx==0.28.1

# YOLOv11 and Computer Vision
ultralytics==8.3.203  # Latest version with YOLOv11 support
torch>=2.0.0
torchvision>=0.15.0
opencv-python==4.11.0.86
pillow==11.3.0
numpy==2.3.3

# Video processing
ffmpeg-python==0.2.0

# AWS services
boto3==1.40.38
ultralytics==8.0.206
numpy==1.24.3
requests==2.31.0
python-multipart==0.0.6
Pillow==10.0.1
scikit-learn==1.3.2
openai==1.3.6
scipy==1.11.4
dataclasses-json==0.6.1
python-socketio==5.10.0
aiohttp==3.9.1
websockets==12.0
scikit-learn==1.5.2
scipy==1.14.1
pandas==2.2.3

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Background tasks and queuing
celery==5.4.0
redis==5.2.0

# OpenAI integration
openai==1.55.1

# Additional utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==24.1.0
websockets==13.1